@@ -26,31 +26,37 @@
     "vue-chat-bot"
   ],
   "devDependencies": {
-    "@commitlint/cli": "^7.6.1",
-    "@commitlint/config-conventional": "^7.6.0",
-    "@types/jest": "^24.0.11",
-    "@vue/cli-plugin-babel": "^4.0.0",
-    "@vue/cli-plugin-eslint": "^4.0.0",
-    "@vue/cli-plugin-unit-jest": "^4.0.0",
-    "@vue/cli-service": "^4.0.0",
-    "@vue/eslint-config-standard": "^4.0.0",
-    "@vue/test-utils": "1.0.0-beta.29",
-    "babel-eslint": "^10.0.3",
-    "eslint": "^5.16.0",
-    "eslint-plugin-vue": "^5.0.0",
-    "standard-version": "^8.0.1"
+    "@babel/core": "^7.22.5",
+    "@babel/eslint-parser": "^7.22.5",
+    "@commitlint/cli": "^17.6.6",
+    "@commitlint/config-conventional": "^17.6.6",
+    "@types/jest": "^29.5.2",
+    "@vue/cli-plugin-babel": "^5.0.8",
+    "@vue/cli-plugin-eslint": "^5.0.8",
+    "@vue/cli-plugin-unit-jest": "^5.0.8",
+    "@vue/cli-service": "^5.0.8",
+    "@vue/compiler-sfc": "^3.3.4",
+    "@vue/eslint-config-standard": "^8.0.1",
+    "@vue/test-utils": "^2.4.0",
+    "@vue/vue2-jest": "^29.2.4",
+    "eslint": "^8.44.0",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-promise": "^6.1.1",
+    "eslint-plugin-vue": "^9.15.1",
+    "jest": "^29.5.0",
+    "standard-version": "^9.5.0"
   },
   "dependencies": {
-    "bootstrap": "^4.6.2",
-    "core-js": "^3.3.2",
+    "bootstrap": "^5.3.0",
+    "core-js": "^3.31.0",
     "document-register-element": "^1.14.10",
-    "sass": "^1.19.0",
-    "sass-loader": "^8.0.2",
-    "vue": "^2.6.10",
+    "mitt": "^3.0.0",
+    "sass": "^1.63.6",
+    "sass-loader": "^13.3.2",
+    "vue": "^3.3.4",
     "vue-custom-element": "^3.3.0",
-    "vue-loader": "^15.8.3",
-    "vue-svg-loader": "^0.15.0",
-    "vue-template-compiler": "^2.6.10"
+    "vue-loader": "^17.2.2"
   },
   "repository": {
     "type": "git",
