@@ -1,8 +1,5 @@
-import Vue from 'vue'
-import App from './App.vue'
+import { createApp } from 'vue';
+import App from './App.vue';
 
-Vue.config.productionTip = false
-
-new Vue({
-  render: (h) => h(App)
-}).$mount('#app')
+// Create and mount Vue 3 app
+createApp(App).mount('#app');
