@@ -1,7 +1,8 @@
 <template>
   <div class="qkb-msg-bubble-component qkb-msg-bubble-component--single-text">
-    <div class="qkb-msg-bubble-component__text" v-if="mainData.type === 'text'">{{ mainData.text }}</div>
-    <div class="qkb-msg-bubble-component__text" v-if="['html', 'button'].includes(mainData.type)" v-html="mainData.text"></div>
+    <div v-if="mainData.type === 'text'" class="qkb-msg-bubble-component__text">{{ mainData.text }}</div>
+    <div v-if="['html', 'button'].includes(mainData.type)" class="qkb-msg-bubble-component__text"
+         v-html="mainData.text"></div>
   </div>
 </template>
 
@@ -9,8 +10,8 @@
 export default {
   props: {
     mainData: {
-      type: Object
-    }
-  }
+      type: Object,
+    },
+  },
 }
 </script>
