@@ -1,4 +1,11 @@
-import Vue from 'vue'
-const EventBus = new Vue()
+// import Vue from 'vue'
+//
+// const EventBus = new Vue()
+//
+// export default EventBus
 
-export default EventBus
+import mitt from "mitt";
+
+const EventBus = mitt();
+
+export default EventBus;
